Move bulk subscriber code to another file under runtime package
